refactor(ClientTUI): make the commands atomic
Make all commands atomic, hence does not send parts of them, for example color + name of the color, instead just write color followed by name of the color on the same line.
In this way is possible to synchronize all the responses to such commands without blocking the updates, including the heartbeat, on the user's input.