refactor: change the way heartbeat is computed
Define the delta as the distance between today (System.currentTimeMillis()) and the last received data from the client. This distance is in milliseconds, if too high the heartbeat will notify a connection lost. The use of a double instead of an integer makes possible to reset the delta something that previously would have been possible only by if the server's thread somehow stops which rarely happens.