feat: add classes to enable raw mode in terminal
Add abstract class Terminal  to provide the methods to enable and disable the raw mode along with the ANSI code related to the supported keys.
The class is inherited by UnixTerminal and WindowsTerminal that provides the methods to enable the raw mode and set the terminal configuration back to its original value.